requirements.txt
google-ads==26.0.1
python-dotenv==1.0.0
pandas==2.2.3 
scikit-learn==1.3.2